Use hand-rolled transition component instead of Collapse.
The new BoxWithTransition method creates a custom functional component
with the supplied transition. The resulting FC only takes shown, sx, and
the implicit children prop, so it doesn't have all the bells and
whistles of MUI Box, but Box doesn't have to do much, so it should be
okay for now.